requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncio-throttle==1.0.2
attrs==25.4.0
audioread==3.1.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.124.2
feedparser==6.0.12
ffmpeg==1.4
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
imgkit==1.2.3
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.23.3
mdurl==0.1.2
midi2audio==0.1.1
mido==1.3.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.4.2
numba==0.63.1
numpy==2.2.6
-e git+https://github.com/openagents-org/openagents.git@6cdee8c057a467d4315d05bd7cb4102ee398e56f#egg=openagents
openai==2.9.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pooch==1.8.2
pretty_midi==0.2.11
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyfluidsynth==1.3.4
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
pyparsing==3.2.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-rtmidi==1.5.8
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
sse-starlette==3.0.3
starlette==0.50.0
structlog==25.5.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.3.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
